+++ src/make_queue/tests/models/test_machines.py
@@ -15,12 +15,6 @@ class TestGenericMachine(TestCase):
     def test_status(self):
         user = User.objects.create_user("test")
 
-        CoursePermission.objects.bulk_create([
-            CoursePermission(name='User is authenticated', short_name='AUTH', description='Permission to reserve machines and workstations at MAKE NTNU'),
-            CoursePermission(name='3D printer course', short_name='3DPR', description='Permission to use 3D printers at MAKE NTNU'),
-            CoursePermission(name='Raise3D course', short_name='R3DP', description='Permission to use the Raise3D printer at MAKE NTNU'),
-            CoursePermission(name='SLA course', short_name='SLAP', description='Permission to use the SLA printer at MAKE NTNU'),
-        ])
         permissions = CoursePermission.objects.all()
 
         Printer3DCourse.objects.create(name="Test", username="test", user=user, date=timezone.localdate())